DBZ-8352 Split offset storage tests
