+++ systemtests/src/test/java/io/debezium/operator/systemtests/ConfigMapOffsetStorageTest.java
@@ -7,7 +7,6 @@ package io.debezium.operator.systemtests;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import java.io.IOException;
 import java.util.Base64;
 
 import org.junit.jupiter.api.Test;
@@ -18,62 +17,15 @@ import io.debezium.operator.api.model.DebeziumServer;
 import io.debezium.operator.api.model.source.Offset;
 import io.debezium.operator.api.model.source.OffsetBuilder;
 import io.debezium.operator.systemtests.resources.NamespaceHolder;
-import io.debezium.operator.systemtests.resources.dmt.DmtClient;
 import io.debezium.operator.systemtests.resources.operator.DebeziumOperatorBundleResource;
 import io.debezium.operator.systemtests.resources.server.DebeziumServerGenerator;
-import io.debezium.operator.systemtests.resources.sinks.RedisResource;
 import io.fabric8.kubernetes.api.model.ConfigMap;
 import io.fabric8.kubernetes.api.model.ConfigMapBuilder;
 import io.fabric8.kubernetes.api.model.ObjectMetaBuilder;
-import io.fabric8.kubernetes.client.LocalPortForward;
 import io.skodjob.testframe.resources.KubeResourceManager;
 
-public class OffsetStorageTest extends TestBase {
+public class ConfigMapOffsetStorageTest extends TestBase {
     private final Logger logger = LoggerFactory.getLogger(this.getClass());
-
-    @Test
-    void testRedisOffsetStorage() {
-        String namespace = NamespaceHolder.INSTANCE.getCurrentNamespace();
-        DebeziumOperatorBundleResource operatorBundleResource = new DebeziumOperatorBundleResource();
-        operatorBundleResource.configureAsDefault(namespace);
-        logger.info("Deploying Operator");
-        operatorBundleResource.deploy();
-        logger.info("Deploying Debezium Server");
-        DebeziumServer server = DebeziumServerGenerator.generateDefaultMysqlToRedis(namespace);
-
-        Offset offset = new OffsetBuilder()
-                .withNewRedis()
-                .withAddress(RedisResource.getDefaultRedisAddress())
-                .endRedis()
-                .withFlushMs(10)
-                .build();
-        server.getSpec().getSource().setOffset(offset);
-
-        KubeResourceManager.getInstance().createResourceWithWait(server);
-        assertStreamingWorks();
-
-        try (LocalPortForward lcp = dmtResource.portForward(portForwardPort, namespace)) {
-            String redis_offset = DmtClient.readRedisOffsets(portForwardHost, portForwardPort);
-            assertThat(redis_offset).contains("file");
-            assertThat(redis_offset).contains("pos");
-        }
-        catch (IOException e) {
-            throw new RuntimeException(e);
-        }
-
-        server.getSpec().getSource().getOffset().getRedis().setKey("metadata:debezium_n:offsets");
-        KubeResourceManager.getInstance().createOrUpdateResourceWithWait(server);
-        assertStreamingWorks(10, 20);
-
-        try (LocalPortForward lcp = dmtResource.portForward(portForwardPort, namespace)) {
-            String redis_offset = DmtClient.readRedisOffsets(portForwardHost, portForwardPort, "metadata:debezium_n:offsets");
-            assertThat(redis_offset).contains("file");
-            assertThat(redis_offset).contains("pos");
-        }
-        catch (IOException e) {
-            throw new RuntimeException(e);
-        }
-    }
 
     @Test
     void testConfigMapOffsetStorage() {
